tools/run_tests/xds_test_driver/requirements.txt
Mako~=1.1
PyYAML~=5.3
absl-py~=0.11
dataclasses~=0.8
google-api-python-client~=1.12
grpcio
grpcio-tools
grpcio-channelz~=1.34
kubernetes~=12.0
retrying~=1.3
